project/requirements.txt
Adafruit-Blinka==8.12.0
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-neopixel==6.3.8
adafruit-circuitpython-pixelbuf==2.0.0
adafruit-circuitpython-requests==1.12.12
adafruit-circuitpython-typing==1.8.3
Adafruit-PlatformDetect==3.39.0
Adafruit-PureIO==1.1.9
autopep8==2.0.1
Faker==16.6.0
mfrc522==0.0.7
paho-mqtt==1.6.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyftdi==0.54.0
pyserial==3.5
python-dateutil==2.8.2
pyusb==1.2.1
RPi.GPIO==0.7.1
six==1.16.0
spidev==3.6
tomli==2.0.1
typing_extensions==4.4.0
